moved add-event.js to admin-routes.js
the idea is to have all admin related pages behind one route
