+++ views/event-form.pug
@@ -14,7 +14,7 @@ block content
     // This block contains the form used by admins to 
     // submit events to the database 
 
-    form(action="/add-event" enctype='multipart/form-data' method="POST").border.border-light.p-5
+    form(action="/admin/new-event/submit" enctype='multipart/form-data' method="POST").border.border-light.p-5
         p.h4.mb-4.text-center New Event
 
         // Event type selector
