+++ views/layout.pug
@@ -41,9 +41,9 @@ html
               | Home
               span.sr-only (current)
           li.nav-item
-            a.nav-link(href='admin') Admin
+            a.nav-link(href='/admin') Admin
           li.nav-item
-            a.nav-link(href='users') Users
+            a.nav-link(href='/users') Users
           // Dropdown
         //-   li.nav-item.dropdown
         //-     a#navbarDropdownMenuLink.nav-link.dropdown-toggle(data-toggle='dropdown' aria-haspopup='true' aria-expanded='false') Dropdown
